takashi
Time Limit: 2 sec / Memory Limit: 1024 MB

Score :
100 points

Problem Statement
Takahashi and Aoki played
N games. You are given a string
S of length
N, representing the results of these games. Takahashi won the
i-th game if the
i-th character of
S is T, and Aoki won that game if it is A.

The overall winner between Takahashi and Aoki is the one who won more games than the other. If they had the same number of wins, the overall winner is the one who reached that number of wins first. Find the overall winner: Takahashi or Aoki.

Constraints
1≤N≤100
N is an integer.
S is a string of length
N consisting of T and A.